Separate port package, restructure kernel to allow other arches If we want to support other architectures, we need to have separate arch-specific binary packages, and a separate port library package.  This commit achieves that by changing the structure to: - xtask (binary package for building - one for all binaries) - x86-64 (binary package) - port (library package)
Other architectures can be added at the top level.  As there is only one xtask package for all architectures, changes will need to be made to allow isolated building and running of these architectures.

In order to have a separate port package, much of the architecture independent code was moved to the new port package.  Devcons was split between port and x86_64: a trait in port, and the struct implementing the trait in x64_64.  More
of the code can be made portable, but we'd need to separate the instantiation of Console from the print function.  This can be done separately.  The arch-specific code is in a directory called 'arch'.  This name should be changed.

Signed-off-by: Graham MacDonald <grahamamacdonald@gmail.com>